Move remaining utils to lib
