+++ src/core/messenger/viewport.ts
@@ -1,6 +1,6 @@
 import { getViewport } from '../../lib/detect/detect-viewport';
 import type { Viewport } from '../../lib/detect/detect-viewport';
-import fastdom from '../../utils/fastdom-promise';
+import fastdom from '../../lib/fastdom-promise';
 import type { RegisterPersistentListener, RespondProxy } from '../messenger';
 
 type IFrameMapValue = {
