+++ src/init/consentless/dynamic/exclusion-slot.ts
@@ -1,5 +1,5 @@
 import { createAdSlot } from '../../../core/create-ad-slot';
-import fastdom from '../../../utils/fastdom-promise';
+import fastdom from '../../../lib/fastdom-promise';
 import { defineSlot } from '../define-slot';
 
 /**
