+++ src/init/shared/set-adtest-in-labels-cookie.ts
@@ -1,5 +1,5 @@
 import { removeCookie, setCookie } from '@guardian/libs';
-import { getUrlVars } from '../../utils/url';
+import { getUrlVars } from '../../lib/url';
 
 /**
  * Not to be confused with set-adtest-cookie.ts!
