+++ src/insert/mobile-sticky.ts
@@ -1,6 +1,6 @@
 import { createAdSlot } from '../core/create-ad-slot';
+import fastdom from '../lib/fastdom-promise';
 import { shouldIncludeMobileSticky } from '../lib/header-bidding/utils';
-import fastdom from '../utils/fastdom-promise';
 import { fillDynamicAdSlot } from './fill-dynamic-advert-slot';
 
 const createAdWrapperClassic = () => {
