+++ src/insert/spacefinder/liveblog-adverts.ts
@@ -3,7 +3,7 @@ import { adSizes } from '../../core/ad-sizes';
 import { createAdSlot } from '../../core/create-ad-slot';
 import { commercialFeatures } from '../../lib/commercial-features';
 import { getCurrentBreakpoint } from '../../lib/detect/detect-breakpoint';
-import fastdom from '../../utils/fastdom-promise';
+import fastdom from '../../lib/fastdom-promise';
 import { fillDynamicAdSlot } from '../fill-dynamic-advert-slot';
 import { spaceFiller } from './space-filler';
 import type {
