+++ src/lib/fastdom-promise.ts
